requirements.txt
cachetools==5.5.0
et_xmlfile==2.0.0
gm==3.0.180
numpy==1.26.4
openpyxl==3.1.5
packaging==26.0
pandas==1.5.3
protobuf==3.20.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
typing==3.7.4.3
tzdata==2023.3
